v1.0.3: Remove onStateChange, update state less unnecessarily
onStateChange proved to be too complex to implement for the time being.
We wanted to support a deeply nested proxy, and report back on each
key that was changed in the state, so we could support the API:
`registerOnStateChange(callback, 'blockList')`, to only catch
updates to the value of blockList, which is an object.
Implementing this perfectly proved to be very difficult, and is likely
overkill for the issue we're trying to solve:
Identifying when a friend is unfriended and updating the friends list UI.